pyproject.toml
[tool.poetry]
name = "bdov1"
version = "0.1.0"
description = ""
authors = ["Zhenya1975 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
gunicorn = "^20.1.0"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
pytz = "^2022.1"
pandas = "^1.4.3"
openpyxl = "^3.0.10"
shortuuid = "^1.0.9"
pyexcel = "^0.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
